rework import

+++ src/common/save/ImportExport.java
@@ -3,8 +3,11 @@ package common.save;
 import api.utils.Utility;
 import common.hud.EnigmaAlert;
 import common.hud.EnigmaLabel;
+import common.hud.EnigmaOptionPane;
 import common.hud.EnigmaWindow;
 import data.config.Config;
+import editor.EditorLauncher;
+import game.EnigmaGame;
 
 import javax.swing.*;
 import java.awt.*;
@@ -37,6 +40,7 @@ public class ImportExport {
      */
     private static ArrayList<boolean[]> toBytes(String s){
         ArrayList<boolean[]> bytes = new ArrayList<>();
+        s += STRING_END;
 
         for(char c : s.toCharArray()){
             String sByte = Integer.toBinaryString(c);
@@ -89,23 +93,33 @@ public class ImportExport {
      * Exporte une map
      * @param mapName Nom de la map
      * @param exportPath Chemin où créer le fichier exporté
-     * @throws IllegalStateException En case d'erreur de lecture ou d'écriture
+     * @throws IllegalStateException En cas d'erreur de lecture ou d'écriture
+     * @throws IllegalStateException Si l'export est annulé
      */
     public static void exportMap(String mapName, String exportPath) throws IOException {
-        DataOutputStream writer = new DataOutputStream(new FileOutputStream(exportPath + mapName + Config.EXPORT_EXTENSION));
+        File file = new File(exportPath + mapName + Config.MAP_EXPORT_EXTENSION);
+
+        if(file.exists()) {
+            if(!EnigmaOptionPane.showConfirmDialog(EditorLauncher.getInstance().getWindow(),
+                    new Dimension(600,250),
+                    "Un fichier nommé \"" + mapName + Config.MAP_EXPORT_EXTENSION + "\" existe déjà, remplacer?")){
+                throw new IllegalStateException("Export annulé");
+            }
+        }
+
+        DataOutputStream writer = new DataOutputStream(new FileOutputStream(file));
 
         try {
-            String[] toWrite = new String[5];
+            String[] toWrite = new String[4];
             toWrite[0] = mapName;
-            toWrite[1] = Utility.readFile(Config.GAME_DATA_FOLDER + mapName + Config.DATA_EXTENSION);
-            toWrite[2] = Utility.readFile(Config.MAP_DATA_FOLDER + mapName + Config.DATA_EXTENSION);
-            toWrite[3] = "";
-            toWrite[4] = Utility.readFile(Config.MAP_FOLDER + mapName + Config.ENIGMA_EXTENSION);
+            toWrite[1] = Utility.readFile(Config.MAP_DATA_FOLDER + mapName + Config.DATA_EXTENSION);
+            toWrite[2] = Utility.readFile(Config.MAP_FOLDER + mapName + Config.MAP_EXTENSION);
+            toWrite[3] = Utility.readFile(Config.MAP_FOLDER + mapName + Config.ENIGMA_EXTENSION);
 
             //Ecriture
             for(String s : toWrite)
-                for(boolean[] tab : ImportExport.toBytes(s + STRING_END))
-                    for(boolean b : tab)
+                for (boolean[] tab : ImportExport.toBytes(s))
+                    for (boolean b : tab)
                         writer.writeBoolean(b);
 
             writer.close();
@@ -119,22 +133,34 @@ public class ImportExport {
      * Exporte une map
      * @param mapName Nom de la map
      * @param exportPath Chemin où créer le fichier exporté
-     * @throws IOException En case d'erreur de lecture ou d'écriture
+     * @throws IOException En cas d'erreur de lecture ou d'écriture
+     * @throws IllegalStateException Si l'export est annulé
      */
-    public static void exportGame(String mapName, String exportPath) throws IOException {
-        DataOutputStream writer = new DataOutputStream(new FileOutputStream(exportPath + mapName + Config.EXPORT_EXTENSION));
+    public static void exportGame(String mapName, String gameName, String exportPath) throws IOException {
+        File file = new File(exportPath + mapName + Config.GAME_EXPORT_EXTENSION);
+
+        if(file.exists()) {
+            if(!EnigmaOptionPane.showConfirmDialog(EditorLauncher.getInstance().getWindow(),
+                    new Dimension(600,250),
+                    "Un fichier nommé \"" + mapName + Config.GAME_EXPORT_EXTENSION + "\" existe déjà, remplacer?")){
+                throw new IllegalStateException("Export annulé");
+            }
+        }
+
+        DataOutputStream writer = new DataOutputStream(new FileOutputStream(file));
 
         try {
-            String[] toWrite = new String[5];
+            String[] toWrite = new String[6];
             toWrite[0] = mapName;
-            toWrite[1] = Utility.readFile(Config.GAME_DATA_FOLDER + mapName + Config.DATA_EXTENSION);
+            toWrite[1] = gameName;
             toWrite[2] = Utility.readFile(Config.MAP_DATA_FOLDER + mapName + Config.DATA_EXTENSION);
-            toWrite[3] = Utility.readFile(Config.MAP_FOLDER + mapName + Config.MAP_EXTENSION);
-            toWrite[4] = Utility.readFile(Config.MAP_FOLDER + mapName + Config.ENIGMA_EXTENSION);
+            toWrite[3] = Utility.readFile(Config.GAME_DATA_FOLDER + gameName + Config.DATA_EXTENSION);
+            toWrite[4] = Utility.readFile(Config.MAP_FOLDER + mapName + Config.MAP_EXTENSION);
+            toWrite[5] = Utility.readFile(Config.MAP_FOLDER + mapName + Config.ENIGMA_EXTENSION);
 
             //Ecriture
             for(String s : toWrite)
-                for(boolean[] tab : ImportExport.toBytes(s + STRING_END))
+                for(boolean[] tab : ImportExport.toBytes(s))
                     for(boolean b : tab)
                         writer.writeBoolean(b);
 
@@ -150,12 +176,11 @@ public class ImportExport {
      * @param importPath Chemin du fichier à importer
      * @throws IOException En case d'erreur de lecture ou d'écriture
      * @throws IllegalArgumentException Si le fichier n'est pas un .enigma
-     * @throws IllegalStateException Si une donnée est corrompue
-     * @throws NumberFormatException Si un numéro de ligne est corrompu
+     * @throws IllegalStateException Si l'import est annulé
      */
-    public static void importGameOrMap(String importPath) throws IOException {
-        if(!importPath.endsWith(Config.EXPORT_EXTENSION))
-            throw new IllegalArgumentException("Ce n'est pas un .enigma!");
+    public static void importMap(String importPath) throws IOException {
+        if(!importPath.endsWith(Config.MAP_EXPORT_EXTENSION))
+            throw new IllegalArgumentException("Ce n'est pas un " + Config.MAP_EXPORT_EXTENSION);
 
         DataInputStream reader = new DataInputStream(new FileInputStream(importPath));
         BufferedWriter writer;
@@ -163,18 +188,84 @@ public class ImportExport {
             //Récupération du nom
             String mapName = ImportExport.readToString(reader);
 
+            for(String s : Utility.getAllMapName()) {
+                if (s.equals(mapName)){
+                    if(!EnigmaOptionPane.showConfirmDialog(EditorLauncher.getInstance().getWindow(),
+                            new Dimension(600,250),
+                            "Une map nommée \"" + mapName + "\" existe déjà, remplacer?")){
+                        throw new IllegalStateException("Import annulé");
+                    }
+                }
+            }
+
             //Récupération des données de la map
             writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(Config.MAP_DATA_FOLDER + mapName + Config.DATA_EXTENSION)));
             writer.write(ImportExport.readToString(reader));
             writer.close();
 
-            //Récupération des données du jeu si il y en a
-            String game;
-            if((game = ImportExport.readToString(reader)).length() > 0) {
-                writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(Config.GAME_DATA_FOLDER + mapName + Config.DATA_EXTENSION)));
-                writer.write(game);
-                writer.close();
+            //Récupération de la map
+            writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(Config.MAP_FOLDER + mapName + Config.MAP_EXTENSION)));
+            writer.write(ImportExport.readToString(reader));
+            writer.close();
+
+            //Récupération des énigmes
+            writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(Config.MAP_FOLDER + mapName + Config.ENIGMA_EXTENSION)));
+            writer.write(ImportExport.readToString(reader));
+            writer.close();
+
+        }catch (EOFException e){
+            reader.close();
+        }
+    }
+
+    /**
+     * Importe une partie
+     * @param importPath Chemin du fichier à importer
+     * @throws IOException En case d'erreur de lecture ou d'écriture
+     * @throws IllegalArgumentException Si le fichier n'est pas un .enigma
+     * @throws IllegalStateException Si l'import est annulé
+     */
+    public static void importGame(String importPath) throws IOException {
+        if(!importPath.endsWith(Config.GAME_EXPORT_EXTENSION))
+            throw new IllegalArgumentException("Ce n'est pas un " + Config.GAME_EXPORT_EXTENSION);
+
+        DataInputStream reader = new DataInputStream(new FileInputStream(importPath));
+        BufferedWriter writer;
+        try {
+            //Récupération du nom
+            String mapName = ImportExport.readToString(reader);
+            String gameName = ImportExport.readToString(reader);
+
+            for(String s : Utility.getAllMapName()) {
+                if (s.equals(mapName)){
+                    if(!EnigmaOptionPane.showConfirmDialog(EditorLauncher.getInstance().getWindow(),
+                            new Dimension(600,250),
+                            "Une map nommée \"" + mapName + "\" existe déjà, remplacer?")){
+                        throw new IllegalStateException("Import annulé");
+                    }
+                }
             }
+
+            if(gameName.length() > 0) {
+                for (String s : Utility.getAllGameName()) {
+                    if (s.equals(gameName)) {
+                        if (!EnigmaOptionPane.showConfirmDialog(EditorLauncher.getInstance().getWindow(),
+                                new Dimension(600, 250),
+                                "Une partie nommée \"" + gameName + "\" existe déjà, remplacer?")) {
+                            throw new IllegalStateException("Import annulé");
+                        }
+                    }
+                }
+            }
+            //Récupération des données de la map
+            writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(Config.MAP_DATA_FOLDER + mapName + Config.DATA_EXTENSION)));
+            writer.write(ImportExport.readToString(reader));
+            writer.close();
+
+            //Récupération des données du jeu
+            writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(Config.GAME_DATA_FOLDER + gameName + Config.DATA_EXTENSION)));
+            writer.write(ImportExport.readToString(reader));
+            writer.close();
 
             //Récupération de la map
             writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(Config.MAP_FOLDER + mapName + Config.MAP_EXTENSION)));
